Working on DDB bindings integration
