+++ Registry.Adapters.DroneDB/Registry.Adapters.DroneDB.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="DDB.Bindings" Version="1.0.1" />
+    <PackageReference Include="DDB.Bindings" Version="1.0.2" />
     <PackageReference Include="GeoJSON.Net" Version="1.2.19" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.8" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.8" />
