+++ Registry.Ports/DroneDB/IDdb.cs
@@ -5,10 +5,14 @@ using Registry.Ports.DroneDB.Models;
 
 namespace Registry.Ports.DroneDB
 {
+    /// <summary>
+    /// Abstracts the drone db operations
+    /// </summary>
     public interface IDdb
     {
-        IEnumerable<DdbInfo> Info(string path);
-        void CreateDatabase(string path);
-        void Remove(string ddbPath, string path);
+        // public DdbObject GetObjectInfo(int id);
+        IEnumerable<DdbEntry> Search(string path);
+        void Add(string path, byte[] data);
+        void Remove(string path);
     }
 }
