--- Registry.Ports/DroneDB/IDdbPackageProvider.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Registry.Ports.DroneDB
-{
-    public interface IDdbPackageProvider
-    {
-        public bool IsDdbReady(bool ignoreVersion = false);
-    }
-}
