+++ Registry.Web.Test/ObjectManagerTest.cs
@@ -71,9 +71,6 @@ namespace Registry.Web.Test
                 File.WriteAllText(Path.Combine(DdbTestDataFolder, "ddbcmd.exe"), string.Empty);
             }
 
-            _settings.DdbPath = DdbTestDataFolder;
-
-
             _ddbFactoryLogger = new Logger<DdbFactory>(LoggerFactory.Create(builder => builder.AddConsole()));
             _objectManagerLogger = new Logger<ObjectsManager>(LoggerFactory.Create(builder => builder.AddConsole()));
         }
@@ -170,7 +167,6 @@ namespace Registry.Web.Test
 
         
         [Test]
-        [Explicit("Does not run in CI")]
         public async Task AddNew_File_FileRes()
         {
