+++ Registry.Web.Test/ShareManagerTest.cs
@@ -95,7 +95,6 @@ namespace Registry.Web.Test
         }
 
         [Test] 
-        [Explicit("Cannot run in CI")]
         public async Task EndToEnd_HappyPath()
         {
             /* INITIALIZATION & SETUP */
@@ -206,7 +205,6 @@ namespace Registry.Web.Test
         }
 
         [Test]
-        [Explicit("Cannot run in CI")]
         public async Task EndToEnd_ShareInit_After_ShareInit()
         {
