+++ Registry.Web/appsettings.Release.json
@@ -17,7 +17,6 @@
       "Password": "password"
     },
     "DdbStoragePath": "./ddbstore",
-    "DdbPath": "",
     "SupportedDdbVersion": {
       "Major": 0,
       "Minor": 9,
