+++ Registry.Web/appsettings.json
@@ -17,7 +17,6 @@
       "Password": "password"
     },
     "DdbStoragePath": "./ddbstore",
-    "DdbPath": "./ddb",
     "SupportedDdbVersion": {
       "Major": 0,
       "Minor": 9,
